Add the method of calculate a square of Parallelogram
