--- HW_06/hw_with_contains/class_Point.py
@@ -1,6 +1,0 @@
-
-
-class Point:
-    def __init__(self, x: int | float, y: int | float):
-        self.x = x
-        self.y = y
